Support setting a vertical divider
